Refactor: 테마뷰모델 생성 및 테마변경 버그 해결
