+++ PJ3T3_Postie/Core/Setting/View/ThemeView.swift
@@ -12,23 +12,13 @@ struct ThemeView: View {
     @AppStorage("selectedLayoutMode") private var selectedLayoutMode: Int = 0
     @AppStorage("isSplitLayout") private var isSplitLayout: Bool = true
     @AppStorage("isThemeGroupButton") private var isThemeGroupButton: Int = 0
+    @StateObject private var viewModel = ThemeViewModel()
     
-    @State private var selectedThemeButton: Int = 0
     @Binding var currentColorPage: Int
     @Binding var isTabGroupButton: Bool
     @Binding var currentGroupPage: Int
     
     var body: some View {
-        let name = ["테마 설정", "나열 변경"]
-        let items = ["포스티 오렌지", "포스티 옐로우", "포스티 그린", "포스티 블루", "포스티 블랙"]
-        let listImages = ["postieListOrange", "postieListYellow", "postieListGreen", "postieListBlue", "postieListBlack"]
-        let groupImages = ["postieGroupOrange", "postieGroupYellow", "postieGroupGreen", "postieGroupBlue", "postieGroupBlack"]
-        
-        let columns: [GridItem] = [
-            GridItem(.flexible()),
-            GridItem(.flexible())
-        ]
-        
         GeometryReader { geometry in
             ZStack {
                 postieColors.backGroundColor
@@ -38,21 +28,21 @@ struct ThemeView: View {
                     HStack(spacing: 10) {
                         ForEach(0...1, id: \.self) { index in
                             Button(action: {
-                                selectedThemeButton = index
+                                viewModel.selectedThemeButton = index
                             }) {
                                 ZStack {
                                     Rectangle()
                                         .foregroundColor(.clear)
                                         .frame(width: 72, height: 30)
-                                        .background(selectedThemeButton == index ? postieColors.tintColor : postieColors.receivedLetterColor)
+                                        .background(viewModel.selectedThemeButton == index ? postieColors.tintColor : postieColors.receivedLetterColor)
                                         .cornerRadius(20)
                                         .shadow(color: Color.postieBlack.opacity(0.1), radius: 3, x: 2, y: 2)
                                     
-                                    Text(name[index])
+                                    Text(viewModel.names[index])
                                         .font(.caption)
-                                        .fontWeight(selectedThemeButton == index ? .bold : .regular)
+                                        .fontWeight(viewModel.selectedThemeButton == index ? .bold : .regular)
                                         .multilineTextAlignment(.center)
-                                        .foregroundColor(selectedThemeButton == index ? postieColors.receivedLetterColor : postieColors.tabBarTintColor)
+                                        .foregroundColor(viewModel.selectedThemeButton == index ? postieColors.receivedLetterColor : postieColors.tabBarTintColor)
                                         .frame(width: 60, alignment: .top)
                                 }
                             }
@@ -60,7 +50,7 @@ struct ThemeView: View {
                         
                         Spacer()
                         
-                        if selectedThemeButton == 0 {
+                        if viewModel.selectedThemeButton == 0 {
                             Menu {
                                 Button(action: {
                                     selectedLayoutMode = 0
@@ -142,10 +132,10 @@ struct ThemeView: View {
                     
                     Spacer()
                     
-                    if selectedThemeButton == 0 {
+                    if viewModel.selectedThemeButton == 0 {
                         if selectedLayoutMode == 0 {
                             TabView(selection: $currentColorPage) {
-                                ForEach(Array(zip(items.indices, items)), id: \.0) { index, item in
+                                ForEach(Array(zip(viewModel.items.indices, viewModel.items)), id: \.0) { index, item in
                                     Button (action: {
                                         isThemeGroupButton = index
                                         currentColorPage = isThemeGroupButton
@@ -159,7 +149,7 @@ struct ThemeView: View {
                                                     .foregroundStyle(isThemeGroupButton == index ? postieColors.tintColor : postieColors.tabBarTintColor)
                                             }
                                             
-                                            Image(isTabGroupButton ? groupImages[index] : listImages[index])
+                                            Image(isTabGroupButton ? viewModel.groupImages[index] : viewModel.listImages[index])
                                                 .resizable()
                                                 .modifier(CustomImageModifier(height: geometry.size.height))
                                                 .border(isThemeGroupButton == index ? postieColors.tintColor : postieColors.tintColor.opacity(0))
@@ -177,14 +167,14 @@ struct ThemeView: View {
                             }
                         } else if selectedLayoutMode == 1 {
                             ScrollView {
-                                LazyVGrid(columns: columns, spacing: 10) {
-                                    ForEach(Array(zip(items.indices, items)), id: \.0) { index, item in
+                                LazyVGrid(columns: viewModel.columns, spacing: 10) {
+                                    ForEach(Array(zip(viewModel.items.indices, viewModel.items)), id: \.0) { index, item in
                                         Button (action: {
                                             isThemeGroupButton = index
                                             ThemeManager.shared.updateTheme(index: index)
                                         }) {
                                             VStack {
-                                                Image(isTabGroupButton ? groupImages[index] : listImages[index])
+                                                Image(isTabGroupButton ? viewModel.groupImages[index] : viewModel.listImages[index])
                                                     .resizable()
                                                     .scaledToFit()
                                                     .frame(height: 200)
@@ -206,14 +196,14 @@ struct ThemeView: View {
                             }
                         } else {
                             ScrollView {
-                                ForEach(Array(zip(items.indices, items)), id: \.0) { index, item in
+                                ForEach(Array(zip(viewModel.items.indices, viewModel.items)), id: \.0) { index, item in
                                     Button (action: {
                                         isThemeGroupButton = index
                                         ThemeManager.shared.updateTheme(index: index)
                                     }) {
                                         VStack {
                                             HStack {
-                                                Image(isTabGroupButton ? groupImages[index] : listImages[index])
+                                                Image(isTabGroupButton ? viewModel.groupImages[index] : viewModel.listImages[index])
                                                     .resizable()
                                                     .scaledToFit()
                                                     .frame(height: 70)
@@ -249,7 +239,7 @@ struct ThemeView: View {
                                                 .foregroundStyle(isTabGroupButton ? postieColors.tintColor : postieColors.tabBarTintColor)
                                         }
                                         
-                                        Image("postieGroup\(stringFromNumber(isThemeGroupButton))")
+                                        Image("postieGroup\(viewModel.stringFromNumber(isThemeGroupButton))")
                                             .resizable()
                                             .modifier(CustomImageModifier(height: geometry.size.height))
                                             .border(isTabGroupButton ? postieColors.tintColor : postieColors.tintColor.opacity(0))
@@ -271,7 +261,7 @@ struct ThemeView: View {
                                                 .foregroundStyle(!isTabGroupButton ? postieColors.tintColor : postieColors.tabBarTintColor)
                                         }
                                         
-                                        Image("postieList\(stringFromNumber(isThemeGroupButton))")
+                                        Image("postieList\(viewModel.stringFromNumber(isThemeGroupButton))")
                                             .resizable()
                                             .modifier(CustomImageModifier(height: geometry.size.height))
                                             .border(isTabGroupButton ? postieColors.tintColor.opacity(0) : postieColors.tintColor)
@@ -293,7 +283,7 @@ struct ThemeView: View {
                                         isTabGroupButton = true
                                     }) {
                                         HStack {
-                                            Image("postieGroup\(stringFromNumber(isThemeGroupButton))")
+                                            Image("postieGroup\(viewModel.stringFromNumber(isThemeGroupButton))")
                                                 .resizable()
                                                 .scaledToFit()
                                                 .frame(height: 70)
@@ -314,7 +304,7 @@ struct ThemeView: View {
                                         isTabGroupButton = false
                                     }) {
                                         HStack {
-                                            Image("postieList\(stringFromNumber(isThemeGroupButton))")
+                                            Image("postieList\(viewModel.stringFromNumber(isThemeGroupButton))")
                                                 .resizable()
                                                 .scaledToFit()
                                                 .frame(height: 70)
@@ -364,21 +354,6 @@ struct CustomImageModifier: ViewModifier {
 
 func saveToUserDefaults<T>(value: T, key: String) {
     UserDefaults.standard.set(value, forKey: key)
-}
-
-private func stringFromNumber(_ number: Int) -> String {
-    switch number {
-    case 1:
-        return "Yellow"
-    case 2:
-        return "Green"
-    case 3:
-        return "Blue"
-    case 4:
-        return "Black"
-    default:
-        return "Orange"
-    }
 }
 
 //#Preview {
